Moved reducers to their own directory, added state to Health Points component
