+++ src/Stats/HealthPoints.js
@@ -1,18 +1,28 @@
 import './HealthPoints.css';
+import { connect } from 'react-redux';
 
-export default function HealthPoints(props){
+function HealthPoints(props){
+    console.log('------->props', props);
     return(
         <div className='health-points-container'>
             <div className='health-points-circle'>
                 <div className='current-health-points-container'>
                     <span className='health-points-indicator'>Current</span>
-                    <span>150 HP</span>
+                    <span>{props.stats.currentHealthPoints} HP</span>
                 </div>
                 <div className='total-health-points-container'>
-                    <span>200 HP</span>
+                    <span>{props.stats.totalHealthPoints} HP</span>
                     <span className='health-points-indicator'>Total</span>
                 </div>
             </div>
         </div>
     )
 }
+
+function mapStateToProps(state) {
+    return {
+        stats: state.stats
+    };
+}
+
+export default connect(mapStateToProps)(HealthPoints);
