+++ src/redux/reducers/rootReducer.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
 import toggleReducer from './toggleReducer';
+import statsReducer from './statsReducer';
 
 const rootReducer = combineReducers({
-    toggles: toggleReducer
+    toggles: toggleReducer,
+    stats: statsReducer
 });
 
 export default rootReducer;
